// types/generated/ContentQueryApiV2.ts
export type ImagesV2 = Array<ImageV2>;
export type ExperienceContentV2 = {
  experience_id: ExperienceId;
  supplier_id: SupplierId;
  language_code: LanguageCode;
  experience_translation?: ExperienceTranslation;
  option_translations?: OptionTranslationArray;
  custom_highlights?: Array<CustomHighlights>;
  custom_included?: Array<CustomIncluded>;
  custom_non_included?: Array<CustomNonIncluded>;
  custom_important_information?: Array<CustomImportantInformation>;
  location?: ExperienceLocation;
};
export type DistributionContentElement = DistributionContentSearchHit | DistributionContent;
export type DistributionContentSearchHit = {
  content: DistributionContent;
  highlight_fields?: {};
};
export type DistributionContent = {
  id?: Id;
  reference_code?: ReferenceCode;
  supplier_id?: SupplierId;
  experience_source?: ExperienceSource;
  global_status?: GlobalStatus;
  experience_media?: ExperienceMedia;
  operational_content?: OperationalContent;
  experience_content?: Array<ExperienceContentV2>;
  functional_content?: FunctionalContent;
};
export type ExperienceMedia = {
  id?: Id;
  images_v2?: ImagesV2;
  flow_code: FlowCode;
  status_code: StatusCode;
  distribution_status?: DistributionStatus;
  distribution_date?: DistributionDate;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type ExperienceTranslation = {
  id: Id;
  name: Name;
  title: Title;
  text1?: Description;
  text2?: ShortDescription;
  text3?: MediumDescription;
  info_voucher?: InfoVoucher;
  meeting_point_details?: MeetingPointDetails;
  automatic_translation?: AutomaticTranslation;
  curation_quality?: CurationQuality;
  flow_code: FlowCode;
  status_code: StatusCode;
  distribution_status?: DistributionStatus;
  distribution_date?: DistributionDate;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type OptionTranslationArray = Array<OptionTranslation>;
export type OptionTranslation = {
  id: Id;
  code: OptionCode;
  description: OptionDescription;
  automatic_translation?: AutomaticTranslation;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type RawElement = RawSearchHit | Raw;
export type RawSearchHit = { content: Raw; highlight_fields?: {} };
export type Raw = {
  id?: Id;
  reference_code?: ReferenceCode;
  experience_id?: ExperienceId;
  experience_source?: ExperienceSource;
  functional?: FunctionalRaw;
  commercial: CommercialRaw;
  operational?: OperationalContent;
  supplier_id?: SupplierId;
  status_code: RawStatusCode;
  flow_code: RawFlowCode;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type FunctionalRaw = {
  asterix_id?: AsterixId;
  core_id?: CoreId;
  external_reference_code?: ExternalReferenceCode;
  options?: Array<Option>;
  highlights?: Array<string>;
  included?: Array<string>;
  non_included?: Array<string>;
  important_information?: Array<string>;
  additional_services?: Array<string>;
  interests?: Array<string>;
  categories?: Array<string>;
  location?: ExperienceLocation;
  markets?: Markets;
  additional_cities?: AdditionalCities;
};
export type CommercialRaw = {
  title: Title;
  description?: Description;
  info_voucher?: InfoVoucher;
  meeting_point_details?: MeetingPointDetails;
  custom_highlights?: Array<RawCustomHighlights>;
  custom_included?: Array<RawCustomIncluded>;
  custom_non_included?: Array<RawCustomNonIncluded>;
  custom_important_information?: Array<RawCustomImportantInformation>;
};
export type OperationalContent = {
  id: Id;
  lastAvailableDateTime?: string;
  cutoffTime?: string;
  calendarType?: string;
  remainingCapacity?: number;
  refreshed?: boolean;
  refreshVersion?: number;
  available: boolean;
};
export type Option = { code: OptionCode; description: string };
export type RawCustomIncluded = {
  code?: string;
  name: string;
  description?: string;
};
export type RawCustomNonIncluded = {
  code?: string;
  name: string;
  description?: string;
};
export type RawCustomHighlights = {
  code?: string;
  name: string;
  description?: string;
};
export type RawCustomImportantInformation = {
  code?: string;
  name: string;
  description?: string;
};
export type CustomIncluded = {
  id: string;
  code?: string;
  name: string;
  description?: string;
  automatic_translation?: AutomaticTranslation;
  visualization_order?: VisualizationOrder;
  flow_code: FlowCode;
  status_code: StatusCode;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type CustomNonIncluded = {
  id: string;
  code?: string;
  name?: string;
  description?: string;
  automatic_translation?: AutomaticTranslation;
  visualization_order?: VisualizationOrder;
  flow_code: FlowCode;
  status_code: StatusCode;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type CustomHighlights = {
  id: string;
  code?: string;
  name: string;
  description?: string;
  automatic_translation?: AutomaticTranslation;
  visualization_order?: VisualizationOrder;
  flow_code: FlowCode;
  status_code: StatusCode;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type CustomImportantInformation = {
  id: string;
  code?: string;
  name: string;
  description?: string;
  automatic_translation?: AutomaticTranslation;
  visualization_order?: VisualizationOrder;
  flow_code: FlowCode;
  status_code: StatusCode;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type FunctionalContent = {
  experience_location?: ExperienceLocation;
  markets?: Markets;
};
export type ExperienceLocation = {
  id?: string;
  option_id?: string;
  latitude?: string;
  longitude?: string;
  altitude?: string;
  address: Address;
  creation_date?: CreationDate;
  updated_date?: UpdatedDate;
};
export type Address = {
  direction?: string;
  city: string;
  postal_code?: string;
  country: string;
};
export type RawStatusCode = "IN_CREATION" | "BEING_CURATED" | "UP_TO_DATE" | "IN_REVIEW" | "SENT_TO_REVIEW";
export type RawFlowCode = "BASE";
export type Id = string;
export type StatusCode = "IN_REVIEW" | "TO_BE_EDIT" | "READY";
export type FlowCode = string;
export type AsterixId = string;
export type CoreId = string;
export type ExternalReferenceCode = string;
export type CreationDate = string;
export type UpdatedDate = string;
export type DistributionDate = string;
export type DistributionStatus = "READY" | "NOT_READY" | "SENT" | "UNPUBLISHED";
export type GlobalStatus = "NOT_READY" | "PUBLISHED" | "UNPUBLISHED";
export type Description = string;
export type ShortDescription = string;
export type MediumDescription = string;
export type Title = string;
export type InfoVoucher = string;
export type MeetingPointDetails = string;
export type AutomaticTranslation = boolean;
export type CurationQuality = boolean;
export type LanguageCode = string;
export type VisualizationOrder = number;
export type SupplierId = string;
export type ReferenceCode = string;
export type ExperienceSource = "SP" | "NOVA" | "SIP" | "ASX" | "NOVANET" | "INTERNAL";
export type ExperienceId = string;
export type FileName = string;
export type PreviewUrl = string;
export type ImageType = string;
export type Name = string;
export type OptionCode = string;
export type OptionDescription = string;
export type Markets = Array<string>;
export type AdditionalCities = Array<string>;